+++ Templates/SharePoint-ADFS/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Add possibility to create a SharePoint Subscription farm running with feature update 22H2.
+* Add option to create a SharePoint Subscription farm running with feature update 22H2.
 * Use a gen2 image for SQL Server VM.
 * Enable LDAPS (LDAP over SSL) on the Active Directory domain.
 * Create a new AD user to run the directory synchronization, and grant it permission "Replicate Directory Changes".